+++ sdk/storage/Azure.Storage.DataMovement.Blobs/src/BlobsStorageResourceProvider.cs
@@ -43,14 +43,6 @@ namespace Azure.Storage.DataMovement.Blobs
         private readonly Func<Uri, ValueTask<AzureSasCredential>> _getAzureSasCredential;
 
         /// <summary>
-        /// Default constructor.
-        /// </summary>
-        public BlobsStorageResourceProvider()
-        {
-            _credentialType = CredentialType.None;
-        }
-
-        /// <summary>
         /// <para>
         /// Constructs this provider to use the given credential when making a new Blob Storage
         /// <see cref="StorageResource"/>.
@@ -356,7 +348,7 @@ namespace Azure.Storage.DataMovement.Blobs
         /// <returns>
         /// The configured storage resource.
         /// </returns>
-        public StorageResource FromClient(
+        public static StorageResource FromClient(
             BlobContainerClient client,
             BlobStorageResourceContainerOptions options = default)
         {
@@ -377,7 +369,7 @@ namespace Azure.Storage.DataMovement.Blobs
         /// <returns>
         /// The configured storage resource.
         /// </returns>
-        public StorageResource FromClient(
+        public static StorageResource FromClient(
             BlockBlobClient client,
             BlockBlobStorageResourceOptions options = default)
         {
@@ -398,7 +390,7 @@ namespace Azure.Storage.DataMovement.Blobs
         /// <returns>
         /// The configured storage resource.
         /// </returns>
-        public StorageResource FromClient(
+        public static StorageResource FromClient(
             PageBlobClient client,
             PageBlobStorageResourceOptions options = default)
         {
@@ -419,7 +411,7 @@ namespace Azure.Storage.DataMovement.Blobs
         /// <returns>
         /// The configured storage resource.
         /// </returns>
-        public StorageResource FromClient(
+        public static StorageResource FromClient(
             AppendBlobClient client,
             AppendBlobStorageResourceOptions options = default)
         {
@@ -472,7 +464,7 @@ namespace Azure.Storage.DataMovement.Blobs
                 {
                     return new BlobStorageResourceContainerOptions()
                     {
-                        BlobDirectoryPrefix = GetPrefix(transferProperties, isSource)
+                        BlobPrefix = GetPrefix(transferProperties, isSource)
                     };
                 }
                 else
@@ -678,9 +670,9 @@ namespace Azure.Storage.DataMovement.Blobs
 
             BlobDestinationCheckpointDetails destinationCheckpointDetails = checkpointDetails as BlobDestinationCheckpointDetails;
 
-            if (null != destinationCheckpointDetails && destinationCheckpointDetails.BlobType?.Value != default)
+            if (null != destinationCheckpointDetails && destinationCheckpointDetails.BlobType != default)
             {
-                return destinationCheckpointDetails.BlobType.Value switch
+                return destinationCheckpointDetails.BlobType switch
                 {
                     BlobType.Block => ResourceType.BlockBlob,
                     BlobType.Page => ResourceType.PageBlob,