+++ sdk/storage/Azure.Storage.DataMovement.Blobs/src/PageBlobStorageResource.cs
@@ -317,15 +317,9 @@ namespace Azure.Storage.DataMovement.Blobs
         protected override StorageResourceCheckpointDetails GetDestinationCheckpointDetails()
         {
             return new BlobDestinationCheckpointDetails(
-                blobType: new(BlobType.Page),
-                contentType: _options?.ContentType,
-                contentEncoding: _options?.ContentEncoding,
-                contentLanguage: _options?.ContentLanguage,
-                contentDisposition: _options?.ContentDisposition,
-                cacheControl: _options?.CacheControl,
-                accessTier: _options?.AccessTier,
-                metadata: _options?.Metadata,
-                tags: default);
+                isBlobTypeSet: true,
+                blobType: BlobType.Page,
+                blobOptions: _options);
         }
 
         // no-op for get permissions