+++ sdk/storage/Azure.Storage.DataMovement.Blobs/tests/ProgressHandlerTests.cs
@@ -301,7 +301,7 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
                 10 /* fileCount */,
                 transferManagerOptions: transferManagerOptions,
                 transferOptions: transferOptions,
-                waitTime: 30);
+                waitTime: 90);
         }
 
         [LiveOnly]
@@ -317,11 +317,10 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
             await PopulateTestContainer(source.Container);
 
             BlobsStorageResourceProvider blobProvider = new(TestEnvironment.Credential);
-            LocalFilesStorageResourceProvider localProvider = new();
 
             TransferManagerOptions transferManagerOptions = new()
             {
-                ResumeProviders = [blobProvider, localProvider]
+                ProvidersForResuming = [blobProvider]
             };
             TransferManager transferManager = new(transferManagerOptions);
 
@@ -339,7 +338,7 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
             // Act - Start transfer
             TransferOperation transfer = await transferManager.StartTransferAsync(
                 await blobProvider.FromContainerAsync(source.Container.Uri),
-                localProvider.FromDirectory(destination.DirectoryPath),
+                LocalFilesStorageResourceProvider.FromDirectory(destination.DirectoryPath),
                 transferOptions);
 
             // TODO: This can likely be replaced with something better once mocking is in place